+++ app/controllers/reservations_controller.rb
@@ -4,16 +4,18 @@ class ReservationsController < ApplicationController
   before_action :authenticate_user!
 
   def index
-    reservations = Reservation.all
+    reservations = policy_scope(Reservation)
     render json: reservations
   end
 
   def show
     reservation = Reservation.find(params[:id])
+    authorize reservation
     render json: reservation
   end
 
   def create
+    authorize Reservation
     reservation_db = Reservation.joins(:screening)
     reservation = reservation_db.new(reservation_params)
     if reservation.save
@@ -26,6 +28,7 @@ class ReservationsController < ApplicationController
 
   def update
     reservation = Reservation.find(params[:id])
+    authorize reservation
     if reservation.update(reservation_params)
       render json: { success: 'Update successful' }
     else
@@ -35,6 +38,7 @@ class ReservationsController < ApplicationController
 
   def delete
     reservation = Reservation.find(params[:id])
+    authorize reservation
     if reservation.destroy
       render json: { success: 'Delete successful' }
     else
